+++ src/bl4ckscor3/bot/bl4ckb0t/misc/AGMaps.java
@@ -15,10 +15,12 @@ public class AGMaps implements Runnable
 {
 	/**The map played before the current one*/
 	private String previousMap;
+	private int previousPlayers;
 
 	public AGMaps()
 	{
 		previousMap = "";
+		previousPlayers = 0;
 	}
 
 	@Override
@@ -32,6 +34,7 @@ public class AGMaps implements Runnable
 			Document doc = Jsoup.connect("http://cache.www.gametracker.com/components/html0/?host=csgo.area-community.net:27210&bgColor=333333&fontColor=CCCCCC&titleBgColor=222222&titleColor=FF9900&borderColor=555555&linkColor=FFCC00&borderLinkColor=222222&showMap=0&showCurrPlayers=0&showTopPlayers=0&showBlogs=0&width=240").userAgent("Mozilla").get();
 			String map = doc.select(".info_line_right > a:nth-child(1)").text();
 			String players = doc.select("div.info_line:nth-child(5) > div:nth-child(2)").text().split("/")[0];
+			String channel = Core.bot.getConfig().getString("AGMapsChannel");
 			
 			if(!previousMap.equals(map) && !map.equals(""))
 			{
@@ -40,8 +43,14 @@ public class AGMaps implements Runnable
 				if(!players.equals(""))
 					message += Colors.NORMAL + " with " + Colors.BOLD + Colors.OLIVE + players + " players " + Colors.NORMAL + "online.";
 
-				Utilities.sendMessage("#AGMaps", message);
+				Utilities.sendMessage(channel, message);
 				previousMap = map;
+				previousPlayers = Integer.parseInt(players);
+			}
+			else if(previousPlayers != Integer.parseInt(players) && !players.equals(""))
+			{
+				Utilities.sendMessage(channel, "There are " + Colors.BOLD + Colors.OLIVE + players + " players " + Colors.NORMAL + "online.");
+				previousPlayers = Integer.parseInt(players);
 			}
 		}
 		catch(Exception e)